R Language Group Project
The sleep efficiency dataset is used to statistically analyse the significance difference between the dependent variable (sleep efficiency) and the independent variable. Literature review is included to select the independent variable related to the dependent variable.  The objective of the research would be to provide more insight towards smart watch users regarding their sleep efficiency by developing a model.